[cmd] Refactor admin command
The new command mixes two oher sub-commands: setup and queues.
It will allow to administer the configuration of the
GrimoireLab services.
